requirements.txt
discord.py
requests-async
